# slackbot-openapi-spec-traige .json
{
  "info": {
    "title": "Events API for Watson Assistant",
    "version": "1.0.1"
  },
  "tags": [],
  "servers": [
    {
      "description": "Code Engine deployment",
      "url": "https://{appname}.{projectid}.{region}.codeengine.appdomain.cloud",
      "variables": {
        "appname": {
          "default": "slackbot-backend",
          "description": "application name"
        },
        "projectid": {
          "default": "projectid",
          "description": "the Code Engine project ID"
        },
        "region": {
          "default": "us-south",
          "description": "the deployment region, e.g., us-south"
        }
      }
    },
    {
      "description": "local test",
      "url": "http://127.0.0.1:{port}",
      "variables": {
        "port": {
          "default": "5000",
          "description": "local port to use"
        }
      }
    }
  ],
  "paths": {
    "/": {
      "get": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          }
        },
        "summary": "Greeting",
        "description": "health check"
      }
    },
    "/patients": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "per_page",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 30
            },
            "required": true
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsOut"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          }
        },
        "summary": "all events",
        "description": "Retrieve all event records",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "post": {
        "parameters": [],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          }
        },
        "summary": "Insert a new event record",
        "description": "Insert a new event record with the given attributes. Its new EID is returned.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventIn"
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/patients/gender/{tgender}": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "per_page",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 30
            },
            "required": true
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
          "in": "path",
                  "name": "tgender",
                  "schema": {
                    "type": "string"
                  },
            "required": true
          }

        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsOut"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          }
        },
        "summary": "records by gender",
        "description": "Retrieve all records of the same gender",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/database/recreate": {
      "post": {
        "parameters": [
          {
            "in": "query",
            "name": "confirmation",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          }
        },
        "summary": "Recreate the database schema",
        "description": "Recreate the database schema and insert sample data.\nRequest must be confirmed by passing query parameter.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/patients/eid/{eid}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "eid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Event record by EID",
        "description": "Retrieve a single event record by its EID",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "eid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Delete an event record by EID",
        "description": "Delete a single event record identified by its EID.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/patients/name/{fname}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "fname",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Event record by name",
        "description": "Retrieve a single record by its name",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    }
  },
  "openapi": "3.0.3",
  "components": {
    "schemas": {
      "ValidationError": {
        "properties": {
          "detail": {
            "type": "object",
            "properties": {
              "<location>": {
                "type": "object",
                "properties": {
                  "<field_name>": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "HTTPError": {
        "properties": {
          "detail": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "current": {
            "type": "string",
            "format": "url"
          },
          "first": {
            "type": "string",
            "format": "url"
          },
          "prev": {
            "type": "string",
            "format": "url"
          },
          "pages": {
            "type": "integer"
          },
          "last": {
            "type": "string",
            "format": "url"
          },
          "per_page": {
            "type": "integer"
          },
          "next": {
            "type": "string",
            "format": "url"
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          }
        }
      },
      "EventOut": {
        "type": "object",
        "properties": {
          "fname": {
            "type": "string"
          },
          "identity": {
            "type": "string"
          },
          "cellnum": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "homeaddress": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          }
        }
      },
      "EventsOut": {
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "patients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventOut"
            }
          }
        }
      },
      "EventIn": {
        "type": "object",
        "properties": {
          "fname": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "identity": {
            "type": "string",
            "minLength": 0,
            "maxLength": 13
          },
          "cellnum": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10
          },
          "email": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "homeaddress": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "gender": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10
          }
        }
        
      },
      "EventsIn": {
        "type": "object",
        "properties": {
          "fname": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "identity": {
            "type": "string",
            "minLength": 0,
            "maxLength": 13
          },
          "cellnum": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10
          },
          "email": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "homeaddress": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "gender": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10
          }
        },
        "required": [
          "fname",
          "identity",
          "cellnum",
          "email",
          "homeaddress",
          "gender"
        ]
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "name": "API_TOKEN",
        "in": "header"
      }
    }
  
  }
}